Copying remote commands through SSH rather than SCP
Previously SCP was used to copy generated scripts to the remotes.
Now whatever arguments are used for the SSH, the same will be used
for the DD over the SSH to send these generated scripts.

Finally, this commit will enable any SSH parameters to be used.
As such, the SSH command can be utilized in its full potential.